+++ blog/house-of-leaves.mdx
@@ -1,0 +1,11 @@
+---
+title: "House of Leaves"
+date: "2022-12-08"
+slug: "house-of-leaves"
+---
+
+House of Leaves is the debut novel by American author Mark Z. Danielewski, published in March 2000 by Pantheon Books. A bestseller, it has been translated into a number of languages, and is followed by a companion piece, The Whalestoe Letters.
+
+The plot is centered on a fictional documentary about a family whose house contains a seemingly endless labyrinth. The format and structure of House of Leaves is unconventional, with unusual page layout and style, making it a prime example of ergodic literature. It contains copious footnotes, many of which contain footnotes themselves, including references to fictional books, films or articles. In contrast, some pages contain only a few words or lines of text, arranged in strange ways to mirror the events in the story, often creating both an agoraphobic and a claustrophobic effect. At points, the book must be rotated to be read. The novel is also distinctive for its multiple narrators, who interact with each other in elaborate and disorienting ways.
+
+While some have attempted to describe the book as a horror story, many readers, as well as the author, define the book as a love story. Danielewski expands on this point in an interview: "I had one woman come up to me in a bookstore and say, 'You know, everyone told me it was a horror book, but when I finished it, I realized that it was a love story.' And she's absolutely right. In some ways, genre is a marketing tool." House of Leaves has also been described as a "satire of academic criticism."
