+++ src/pages/blog/{mdx.frontmatter__slug}.js
@@ -1,0 +1,27 @@
+import * as React from 'react'
+import { graphql } from 'gatsby'
+import Layout from '../../components/layout'
+import Seo from '../../components/seo'
+
+const BlogPost = ({ data, children }) => {
+  return (
+    <Layout pageTitle={data.mdx.frontmatter.title}>
+      <p>{data.mdx.frontmatter.date}</p>
+      {children}
+    </Layout>
+  )
+}
+export const query = graphql`
+query ($id: String) {
+  mdx(id: {eq: $id}) {
+    frontmatter {
+      date(formatString: "D MMMM, YYYY")
+      title
+    }
+  }
+}
+`
+
+export const Head = ({ data }) => <Seo title={data.mdx.frontmatter.title} />
+
+export default BlogPost
